+++ tests/integration/components/option-test.js
@@ -27,7 +27,7 @@ module('select-box/option', function (hooks) {
       .exists({ count: 1 }, 'has a data attribute signifying its type');
   });
 
-  test('aria role', async function (assert) {
+  test('role', async function (assert) {
     assert.expect(1);
 
     await render(hbs`<SelectBox::Option />`);